move  api token credential to common
