+++ src/Maestro/Client/src/MaestroApi.cs
@@ -92,7 +92,7 @@ namespace Microsoft.DotNet.Maestro.Client
             // 1. BAR or Entra token that can directly be used to authenticate against Maestro
             if (!string.IsNullOrEmpty(barApiToken))
             {
-                return new MaestroApiTokenCredential(barApiToken!);
+                return new ResolvedCredential(barApiToken!);
             }
 
             barApiBaseUri ??= ProductionBuildAssetRegistryBaseUri;
